Deleted TimeSimulation and max_distance_from_speed in favour of Simulation and run_simulation. Renamed 'array' module to 'arrays' to avoid weird import ambiguities.
