--- simulation/array/__init__.py
@@ -1,2 +1,0 @@
-from simulation.array.base_array import BaseArray
-from simulation.array.basic_array import BasicArray
