+++ simulation/arrays/basic_array.py
@@ -1,4 +1,4 @@
-from simulation.array.base_array import BaseArray
+from simulation.arrays.base_array import BaseArray
 
 
 class BasicArray(BaseArray):
