+++ simulation/main/ExecuteSimulation.py
@@ -1,7 +1,7 @@
 import numpy as np
 import json
 
-from simulation.main.MainSimulation import Simulation, SimulationReturnType
+from simulation.main.Simulation import Simulation, SimulationReturnType
 from simulation.common.simulationState import SimulationState
 from simulation.main.SimulationResult import SimulationResult
 from simulation.config import settings_directory
@@ -12,7 +12,7 @@ Description: Export Simulation data as a SimulationResults object.
 """
 
 
-def GetSimulationData() -> SimulationResult:
+def GetSimulationData(golang=True) -> SimulationResult:
     """
     Returns a SimulationResult object with the purpose of exporting simulation data.
     """
@@ -24,7 +24,7 @@ def GetSimulationData() -> SimulationResult:
 
     return_type = SimulationReturnType.simulation_results
     initialSimulationConditions = SimulationState(args)
-    simulation_model = Simulation(initialSimulationConditions, return_type, race_type="ASC")
+    simulation_model = Simulation(initialSimulationConditions, return_type, race_type="ASC", golang=golang)
 
     optimized = simulation_model.run_model(speed=input_speed, plot_results=False, verbose=False, route_visualization=False,
                                            return_results_object=True)
