+++ simulation/main/__init__.py
@@ -1,3 +1,2 @@
-from simulation.main.MainSimulation import Simulation
+from simulation.main.Simulation import Simulation
 from simulation.main.SimulationResult import SimulationResult
-from simulation.main.TimeSimulation import TimeSimulation
